feat:C端小程序登录&商品浏览

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -103,7 +103,7 @@ public class DishController {
      */
     @GetMapping("/list")
     @ApiOperation("根据分类id查询菜品列表")
-    public Result<List<Dish>> list(Integer categoryId) {
+    public Result<List<Dish>> list(Long categoryId) {
         log.info("根据分类id查询菜品列表：{}", categoryId);
         List<Dish> dishes = dishService.list(categoryId);
         return Result.success(dishes);
